// src/API.ts
export type ModelStringKeyConditionInput = {
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
};

export type ModelRACSensorsFilterInput = {
  CO2?: ModelFloatInput | null,
  MAC?: ModelIntInput | null,
  O2?: ModelFloatInput | null,
  atm_temperature?: ModelFloatInput | null,
  conductivity?: ModelFloatInput | null,
  humidity?: ModelFloatInput | null,
  pH?: ModelFloatInput | null,
  temperature?: ModelFloatInput | null,
  timestamp?: ModelStringInput | null,
  and?: Array< ModelRACSensorsFilterInput | null > | null,
  or?: Array< ModelRACSensorsFilterInput | null > | null,
  not?: ModelRACSensorsFilterInput | null,
};
// ...
export enum ModelSortDirection {
  ASC = "ASC",
  DESC = "DESC",
}


export type ModelRACSensorsConnection = {
  __typename: "ModelRACSensorsConnection",
  items:  Array<RACSensors | null >,
  nextToken?: string | null,
};

export type RACSensors = {
  __typename: "RACSensors",
  CO2?: number | null,
  MAC: number,
  O2?: number | null,
  atm_temperature?: number | null,
  conductivity?: number | null,
  humidity?: number | null,
  pH?: number | null,
  temperature?: number | null,
  timestamp: string,
};

export type UpdateActuationByVariableInput = {
  Min: number,
};

export type ActuationMin = {
  __typename: "ActuationMin",
  Min?: number | null,
  PesudoMin?: number | null,
  Setup: string,
  Variable: string,
};
// ...
export type ActuationMax = {
  __typename: "ActuationMax",
  Max?: number | null,
  PesudoMax?: number | null,
  Setup: string,
  Variable: string,
};
// ...
export type SoilSensors = {
  __typename: "SoilSensors",
  CO2?: number | null,
  MAC: number,
  O2?: number | null,
  atm_temperature?: number | null,
  conductivity?: number | null,
  humidity?: number | null,
  Nitrogen?: number | null,
  Phosphorous?: number | null,
  pH?: number | null,
  temperature?: number | null,
  timestamp: string,
};
// ...
export type PlantRepository = {
  __typename: "PlantRepository",
  Crops: string,
  Name: string,
  nitrogen_upper?: number | null,
  nitrogen_lower?: number | null,
  ec_upper?: number | null,
  ec_lower?: number | null,
  pH_upper?: number | null,
  pH_lower?: number | null,
};
// ...
export type EventTemplate = {
  __typename: "EventTemplate",
  userId?: string | null,
  eventName: string,
  description?: string | null,
  duration: number,
  id: string,
};

export type UpdateEventTemplateInput = {
  userId?: string | null,
  eventName?: string | null,
  description?: string | null,
  duration?: number | null,
  id: string,
};

export type DeleteEventTemplateInput = {
  id: string,
};

export type ModelActuationMaxFilterInput = {
  Max?: ModelFloatInput | null,
  PesudoMax?: ModelFloatInput | null,
  Setup?: ModelStringInput | null,
  Variable?: ModelStringInput | null,
  and?: Array< ModelActuationMaxFilterInput | null > | null,
  or?: Array< ModelActuationMaxFilterInput | null > | null,
  not?: ModelActuationMaxFilterInput | null,
};

export type ModelActuationMaxConnection = {
  __typename: "ModelActuationMaxConnection",
  items:  Array<ActuationMax | null >,
  nextToken?: string | null,
};

export type ModelActuationMinFilterInput = {
  Min?: ModelFloatInput | null,
  PesudoMin?: ModelFloatInput | null,
  Setup?: ModelStringInput | null,
  Variable?: ModelStringInput | null,
  and?: Array< ModelActuationMinFilterInput | null > | null,
  or?: Array< ModelActuationMinFilterInput | null > | null,
  not?: ModelActuationMinFilterInput | null,
};

export type ModelActuationMinConnection = {
  __typename: "ModelActuationMinConnection",
  items:  Array<ActuationMin | null >,
  nextToken?: string | null,
};

export type ModelSoilSensorsFilterInput = {
  CO2?: ModelFloatInput | null,
  MAC?: ModelIntInput | null,
  O2?: ModelFloatInput | null,
  atm_temperature?: ModelFloatInput | null,
  conductivity?: ModelFloatInput | null,
  humidity?: ModelFloatInput | null,
  Nitrogen?: ModelFloatInput | null,
  Phosphorous?: ModelFloatInput | null,
  pH?: ModelFloatInput | null,
  temperature?: ModelFloatInput | null,
  timestamp?: ModelStringInput | null,
  and?: Array< ModelSoilSensorsFilterInput | null > | null,
  or?: Array< ModelSoilSensorsFilterInput | null > | null,
  not?: ModelSoilSensorsFilterInput | null,
};

export type ModelSoilSensorsConnection = {
  __typename: "ModelSoilSensorsConnection",
  items:  Array<SoilSensors | null >,
  nextToken?: string | null,
};

export type ModelPlantRepositoryFilterInput = {
  Crops?: ModelStringInput | null,
  Name?: ModelStringInput | null,
  nitrogen_upper?: ModelFloatInput | null,
  nitrogen_lower?: ModelFloatInput | null,
  ec_upper?: ModelFloatInput | null,
  ec_lower?: ModelFloatInput | null,
  pH_upper?: ModelFloatInput | null,
  pH_lower?: ModelFloatInput | null,
  and?: Array< ModelPlantRepositoryFilterInput | null > | null,
  or?: Array< ModelPlantRepositoryFilterInput | null > | null,
  not?: ModelPlantRepositoryFilterInput | null,
};

export type ModelPlantRepositoryConnection = {
  __typename: "ModelPlantRepositoryConnection",
  items:  Array<PlantRepository | null >,
  nextToken?: string | null,
};

export type ModelEventTemplateFilterInput = {
  userId?: ModelStringInput | null,
  eventName?: ModelStringInput | null,
  description?: ModelStringInput | null,
  duration?: ModelIntInput | null,
  and?: Array< ModelEventTemplateFilterInput | null > | null,
  or?: Array< ModelEventTemplateFilterInput | null > | null,
  not?: ModelEventTemplateFilterInput | null,
};

export type ModelEventTemplateConnection = {
  __typename: "ModelEventTemplateConnection",
  items:  Array<EventTemplate | null >,
  nextToken?: string | null,
};
// ...
export type GetActuationMaxQueryVariables = {
  Setup: string,
  Variable: string,
};

export type GetActuationMaxQuery = {
  getActuationMax?:  {
    __typename: "ActuationMax",
    Max?: number | null,
    PesudoMax?: number | null,
    Setup: string,
    Variable: string,
  } | null,
};

export type ListActuationMaxesQueryVariables = {
  Setup?: string | null,
  Variable?: ModelStringKeyConditionInput | null,
  filter?: ModelActuationMaxFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
  sortDirection?: ModelSortDirection | null,
};

export type ListActuationMaxesQuery = {
  listActuationMaxes?:  {
    __typename: "ModelActuationMaxConnection",
    items:  Array< {
      __typename: "ActuationMax",
      Max?: number | null,
      PesudoMax?: number | null,
      Setup: string,
      Variable: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type GetActuationMinQueryVariables = {
  Setup: string,
  Variable: string,
};

export type GetActuationMinQuery = {
  getActuationMin?:  {
    __typename: "ActuationMin",
    Min?: number | null,
    PesudoMin?: number | null,
    Setup: string,
    Variable: string,
  } | null,
};

export type ListActuationMinsQueryVariables = {
  Setup?: string | null,
  Variable?: ModelStringKeyConditionInput | null,
  filter?: ModelActuationMinFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
  sortDirection?: ModelSortDirection | null,
};

export type ListActuationMinsQuery = {
  listActuationMins?:  {
    __typename: "ModelActuationMinConnection",
    items:  Array< {
      __typename: "ActuationMin",
      Min?: number | null,
      PesudoMin?: number | null,
      Setup: string,
      Variable: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type GetRACSensorsQueryVariables = {
  MAC: number,
  timestamp: string,
};

export type GetRACSensorsQuery = {
  getRACSensors?:  {
    __typename: "RACSensors",
    CO2?: number | null,
    MAC: number,
    O2?: number | null,
    atm_temperature?: number | null,
    conductivity?: number | null,
    humidity?: number | null,
    pH?: number | null,
    temperature?: number | null,
    timestamp: string,
  } | null,
};

export type ListRACSensorsQueryVariables = {
  MAC?: number | null,
  timestamp?: ModelStringKeyConditionInput | null,
  filter?: ModelRACSensorsFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
  sortDirection?: ModelSortDirection | null,
};

export type ListRACSensorsQuery = {
  listRACSensors?:  {
    __typename: "ModelRACSensorsConnection",
    items:  Array< {
      __typename: "RACSensors",
      CO2?: number | null,
      MAC: number,
      O2?: number | null,
      atm_temperature?: number | null,
      conductivity?: number | null,
      humidity?: number | null,
      pH?: number | null,
      temperature?: number | null,
      timestamp: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type GetSoilSensorsQueryVariables = {
  MAC: number,
  timestamp: string,
};

export type GetSoilSensorsQuery = {
  getSoilSensors?:  {
    __typename: "SoilSensors",
    CO2?: number | null,
    MAC: number,
    O2?: number | null,
    atm_temperature?: number | null,
    conductivity?: number | null,
    humidity?: number | null,
    Nitrogen?: number | null,
    Phosphorous?: number | null,
    pH?: number | null,
    temperature?: number | null,
    timestamp: string,
  } | null,
};

export type ListSoilSensorsQueryVariables = {
  MAC?: number | null,
  timestamp?: ModelStringKeyConditionInput | null,
  filter?: ModelSoilSensorsFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
  sortDirection?: ModelSortDirection | null,
};

export type ListSoilSensorsQuery = {
  listSoilSensors?:  {
    __typename: "ModelSoilSensorsConnection",
    items:  Array< {
      __typename: "SoilSensors",
      CO2?: number | null,
      MAC: number,
      O2?: number | null,
      atm_temperature?: number | null,
      conductivity?: number | null,
      humidity?: number | null,
      Nitrogen?: number | null,
      Phosphorous?: number | null,
      pH?: number | null,
      temperature?: number | null,
      timestamp: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type GetPlantRepositoryQueryVariables = {
  Crops: string,
};

export type GetPlantRepositoryQuery = {
  getPlantRepository?:  {
    __typename: "PlantRepository",
    Crops: string,
    Name: string,
    nitrogen_upper?: number | null,
    nitrogen_lower?: number | null,
    ec_upper?: number | null,
    ec_lower?: number | null,
    pH_upper?: number | null,
    pH_lower?: number | null,
  } | null,
};

export type ListPlantRepositoriesQueryVariables = {
  Crops?: string | null,
  filter?: ModelPlantRepositoryFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
  sortDirection?: ModelSortDirection | null,
};

export type ListPlantRepositoriesQuery = {
  listPlantRepositories?:  {
    __typename: "ModelPlantRepositoryConnection",
    items:  Array< {
      __typename: "PlantRepository",
      Crops: string,
      Name: string,
      nitrogen_upper?: number | null,
      nitrogen_lower?: number | null,
      ec_upper?: number | null,
      ec_lower?: number | null,
      pH_upper?: number | null,
      pH_lower?: number | null,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type GetEventTemplateQueryVariables = {
  id: string,
};

export type GetEventTemplateQuery = {
  getEventTemplate?:  {
    __typename: "EventTemplate",
    userId?: string | null,
    eventName: string,
    description?: string | null,
    duration: number,
    id: string,
  } | null,
};

export type ListEventTemplatesQueryVariables = {
  filter?: ModelEventTemplateFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListEventTemplatesQuery = {
  listEventTemplates?:  {
    __typename: "ModelEventTemplateConnection",
    items:  Array< {
      __typename: "EventTemplate",
      userId?: string | null,
      eventName: string,
      description?: string | null,
      duration: number,
      id: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type GetUserEventTemplatesQueryVariables = {
  userId: string,
  sortDirection?: ModelSortDirection | null,
  filter?: ModelEventTemplateFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type GetUserEventTemplatesQuery = {
  getUserEventTemplates?:  {
    __typename: "ModelEventTemplateConnection",
    items:  Array< {
      __typename: "EventTemplate",
      userId?: string | null,
      eventName: string,
      description?: string | null,
      duration: number,
      id: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};